chore: Split fork_gc.c into smaller submodules
